# packages/api-contract/openapi/merchant-web.yaml
openapi: 3.1.0
info:
  title: 豆包获客普通商户网页客户端 API
  version: 0.1.0
  description: 普通商户网页客户端真实 API 契约。带 -mock 后缀的历史路径由服务端明确拒绝，仅为旧版本地 UI 测试保留兼容记录，不得用于业务调用。
servers:
  - url: /api
paths:
...
  /merchant/bootstrap:
    get:
      operationId: getMerchantBootstrap
      responses:
        '200':
          description: 当前商户、品牌和会话上下文
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/dashboard:
    get:
      operationId: getMerchantDashboard
      responses:
        '200':
          description: 商户控制台可追溯汇总
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/profile:
    get:
      operationId: getMerchantProfile
      responses:
        '200':
          description: 当前商户唯一企业事实源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantProfile'
        '401':
          $ref: '#/components/responses/ApiError'
    put:
      operationId: updateMerchantProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: 已保存并创建新的企业资料版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantProfile'
        '401':
          $ref: '#/components/responses/ApiError'
        '422':
          $ref: '#/components/responses/ApiError'
  /merchant/keywords:
    get:
      operationId: listMerchantKeywords
      responses:
        '200':
          description: 未删除的核心关键词
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MerchantKeyword'
        '401':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createMerchantKeyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeywordRequest'
      responses:
        '201':
          description: 已创建关键词
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantKeyword'
        '401':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/keywords/{keywordId}:
    parameters:
      - $ref: '#/components/parameters/KeywordId'
    patch:
      operationId: updateMerchantKeyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateKeywordRequest'
      responses:
        '200':
          description: 已更新关键词状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantKeyword'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deleteMerchantKeyword
      responses:
        '204':
          description: 已逻辑删除关键词
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/keywords/{keywordId}/questions:
    parameters:
      - $ref: '#/components/parameters/KeywordId'
    get:
      operationId: listMerchantQuestions
      responses:
        '200':
          description: 指定关键词下未删除的问题词
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MerchantQuestion'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createMerchantQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionRequest'
      responses:
        '201':
          description: 已手动创建问题词
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantQuestion'
        '401':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/keywords/{keywordId}/questions/expand-mock:
    parameters:
      - $ref: '#/components/parameters/KeywordId'
    post:
      deprecated: true
      operationId: expandMerchantQuestionsMock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockQuestionExpandRequest'
      responses:
        '200':
          description: 历史本地 Mock 路径；服务端拒绝调用。请使用 /expand 创建真实异步任务。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockQuestionExpandResponse'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/keywords/{keywordId}/questions/batch:
    parameters:
      - $ref: '#/components/parameters/KeywordId'
    post:
      operationId: createMerchantQuestionsBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionBatchCreateRequest'
      responses:
        '201':
          description: 已批量创建并去重问题词
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionBatchCreateResponse'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/keywords/{keywordId}/questions/expand:
    parameters:
      - $ref: '#/components/parameters/KeywordId'
    post:
      operationId: createQuestionExpansionTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionExpansionTaskInput'
      responses:
        '201':
          description: 已创建问题词拓展异步任务；仅成功新增的问题词扣 1 算力点
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiGenerationTask'
        '401':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/QuestionId'
    patch:
      operationId: updateMerchantQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionRequest'
      responses:
        '200':
          description: 已更新问题词状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantQuestion'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deleteMerchantQuestion
      responses:
        '204':
          description: 已逻辑删除问题词
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/knowledge-libraries:
    get:
      operationId: listKnowledgeLibraries
      responses:
        '200':
          description: 商户企业信息库
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeLibrary'
        '401':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createKnowledgeLibrary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeLibraryInput'
      responses:
        '201':
          description: 已创建信息库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeLibrary'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/knowledge-libraries/{libraryId}:
    parameters:
      - $ref: '#/components/parameters/KnowledgeLibraryId'
    put:
      operationId: updateKnowledgeLibrary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeLibraryInput'
      responses:
        '200':
          description: 已更新信息库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeLibrary'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deleteKnowledgeLibrary
      responses:
        '204':
          description: 已删除信息库
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/galleries:
    get:
      operationId: listMerchantGalleries
      responses:
        '200':
          description: 商户企业图库
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MerchantGallery'
        '401':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createMerchantGallery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryInput'
      responses:
        '201':
          description: 已创建图库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantGallery'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/galleries/{galleryId}:
    parameters:
      - $ref: '#/components/parameters/GalleryId'
    put:
      operationId: updateMerchantGallery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryInput'
      responses:
        '200':
          description: 已更新图库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantGallery'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deleteMerchantGallery
      responses:
        '204':
          description: 已删除图库及其图片元数据
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/galleries/{galleryId}/images:
    parameters:
      - $ref: '#/components/parameters/GalleryId'
    get:
      operationId: listMerchantGalleryImages
      responses:
        '200':
          description: 图库图片元数据
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GalleryImage'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createMerchantGalleryImageUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryImageInput'
      responses:
        '201':
          description: 已创建 5 分钟有效的 OSS 直传会话；尚未产生图片元数据或空间用量
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryImageUploadTicket'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/galleries/{galleryId}/images/{uploadId}/complete:
    parameters:
      - $ref: '#/components/parameters/GalleryId'
      - $ref: '#/components/parameters/UploadId'
    post:
      operationId: completeMerchantGalleryImageUpload
      responses:
        '201':
          description: 对象已完成校验并加入图库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryImage'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/gallery-images/{imageId}:
    parameters:
      - $ref: '#/components/parameters/GalleryImageId'
    delete:
      operationId: deleteMerchantGalleryImage
      responses:
        '204':
          description: 已删除图片元数据
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/writing-instructions:
    get:
      operationId: listWritingInstructions
      responses:
        '200':
          description: 系统和自定义创作指令
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WritingInstruction'
        '401':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createWritingInstruction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WritingInstructionInput'
      responses:
        '201':
          description: 已创建自定义创作指令
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WritingInstruction'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/writing-instructions/{instructionId}:
    parameters:
      - $ref: '#/components/parameters/WritingInstructionId'
    put:
      operationId: updateWritingInstruction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WritingInstructionInput'
      responses:
        '200':
          description: 已更新自定义创作指令
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WritingInstruction'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deleteWritingInstruction
      responses:
        '204':
          description: 已删除自定义创作指令
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/articles:
    get:
      operationId: listMerchantArticles
      responses:
        '200':
          description: 商户文章列表，包含真实 AI 生成文章和手动文章
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MerchantArticle'
        '401':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createMerchantArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '201':
          description: 已手动创建文章
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantArticle'
        '401':
          $ref: '#/components/responses/ApiError'
        '422':
          $ref: '#/components/responses/ApiError'
  /merchant/articles/create-mock:
    post:
      deprecated: true
      operationId: createMerchantArticlesMock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockArticleCreateRequest'
      responses:
        '201':
          description: 历史本地 Mock 路径；服务端拒绝调用。请使用 /ai-tasks 创建真实异步任务。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockArticleCreateResponse'
        '401':
          $ref: '#/components/responses/ApiError'
        '422':
          $ref: '#/components/responses/ApiError'
  /merchant/ai-tasks:
    get:
      operationId: listMerchantAiTasks
      responses:
        '200':
          description: 商户自己的 AI 任务列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiGenerationTask' }
        '401':
          $ref: '#/components/responses/ApiError'
    post:
      operationId: createMerchantArticleAiTask
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArticleAiTaskInput' }
      responses:
        '201':
          description: 已创建并预占额度的文章创作任务
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiGenerationTask' }
        '401':
          $ref: '#/components/responses/ApiError'
        '422':
          $ref: '#/components/responses/ApiError'
  /merchant/ai-tasks/{taskId}/stop:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: stopMerchantAiTask
      responses:
        '200':
          description: 已请求停止任务
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiGenerationTask' }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/ai-tasks/{taskId}/retryable-questions:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: listMerchantAiTaskRetryableQuestions
      responses:
        '200':
          description: 已结束文章任务中仍未生成、可单篇重试的问题词
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiTaskRetryableQuestion' }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/ai-tasks/{taskId}/retry:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string }
      - name: Idempotency-Key
        in: header
        required: true
        schema: { type: string, minLength: 8, maxLength: 100 }
    post:
      operationId: retryMerchantAiTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiTaskRetryInput' }
      responses:
        '201':
          description: 已创建新的文章重试任务并预占对应额度
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiGenerationTask' }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/articles/{articleId}:
    parameters:
      - $ref: '#/components/parameters/ArticleId'
    put:
      operationId: updateMerchantArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleInput'
      responses:
        '200':
          description: 已更新文章
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantArticle'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /merchant/website:
    get:
      operationId: getMerchantWebsite
      responses:
        '200':
          description: 商户当前网站模板与本地生成状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantWebsite'
        '401':
          $ref: '#/components/responses/ApiError'
    put:
      operationId: updateMerchantWebsite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantWebsiteInput'
      responses:
        '200':
          description: 已保存网站模板选择
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantWebsite'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/website/generate-mock:
    post:
      deprecated: true
      operationId: generateMerchantWebsiteMock
      responses:
        '200':
          description: 历史本地 Mock 路径；服务端拒绝调用。请使用 /website/generate 生成本地静态站。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantWebsite'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/website/generate:
    post:
      operationId: generateMerchantWebsite
      responses:
        '201':
          description: 已完成本地静态站渲染，可通过本机预览地址核验；尚未发布到自定义域名
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantWebsite'
        '401':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
  /merchant/doubao-results:
    get:
      operationId: listMerchantDoubaoResults
      responses:
        '200':
          description: 普通商户只读的最近豆包名称命中检测结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoubaoCheckResult'
        '401':
          $ref: '#/components/responses/ApiError'
  /merchant/media-accounts:
    get:
      operationId: listMerchantMediaAccounts
      responses:
        '200':
          description: 仅包含本地助手回传的非凭证账号状态
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MediaAccount' }
        '401': { $ref: '#/components/responses/ApiError' }
  /merchant/media-accounts/{platform}/connect-mock:
    parameters:
      - $ref: '#/components/parameters/MediaPlatform'
    post:
      deprecated: true
      operationId: requestMediaAccountConnectMock
      responses:
        '200':
          description: 历史本地 Mock 路径；服务端拒绝调用。真实扫码由本地 EXE 助手处理。
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaAccount' }
        '401': { $ref: '#/components/responses/ApiError' }
...
components:
  parameters:
...
    KeywordId:
      name: keywordId
      in: path
      required: true
      schema:
        type: string
    QuestionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ApiError:
      description: 稳定业务错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    BootstrapResponse:
      type: object
      additionalProperties: false
      required: [brand, account]
      properties:
        brand:
          $ref: '#/components/schemas/BrandContext'
        account:
          $ref: '#/components/schemas/MerchantAccount'
...
    DashboardResponse:
      type: object
      additionalProperties: false
      required: [resources, effects, inclusionTrend, workflow, recentTasks, lastCheckedAt]
      properties:
        resources:
          $ref: '#/components/schemas/ResourceSummary'
        effects:
          $ref: '#/components/schemas/EffectSummary'
        inclusionTrend:
          type: array
          items:
            $ref: '#/components/schemas/TrendPoint'
        workflow:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: '#/components/schemas/WorkflowStage'
        recentTasks:
          type: array
          items:
            $ref: '#/components/schemas/RecentTask'
        lastCheckedAt:
          type: string
          format: date-time
          nullable: true
...